+++ espflash/src/cli/mod.rs
@@ -723,10 +723,12 @@ pub fn erase_flash(args: EraseFlashArgs, config: &Config) -> Result<()> {
     let mut flasher = connect(&args.connect_args, config, true, true)?;
     info!("Erasing Flash...");
 
+    let chip = flasher.chip();
+
     flasher.erase_flash()?;
     flasher
         .connection()
-        .reset_after(!args.connect_args.no_stub)?;
+        .reset_after(!args.connect_args.no_stub, chip)?;
 
     info!("Flash has been erased!");
 
@@ -747,6 +749,7 @@ pub fn erase_region(args: EraseRegionArgs, config: &Config) -> Result<()> {
     }
 
     let mut flasher = connect(&args.connect_args, config, true, true)?;
+    let chip = flasher.chip();
 
     info!(
         "Erasing region at 0x{:08x} ({} bytes)",
@@ -756,7 +759,7 @@ pub fn erase_region(args: EraseRegionArgs, config: &Config) -> Result<()> {
     flasher.erase_region(args.address, args.size)?;
     flasher
         .connection()
-        .reset_after(!args.connect_args.no_stub)?;
+        .reset_after(!args.connect_args.no_stub, chip)?;
 
     Ok(())
 }