+++ espflash/src/connection/mod.rs
@@ -33,7 +33,10 @@ use self::{
         UsbJtagSerialReset,
     },
 };
-use crate::error::{ConnectionError, Error, ResultExt, RomError, RomErrorKind};
+use crate::{
+    error::{ConnectionError, Error, ResultExt, RomError, RomErrorKind},
+    targets::Chip,
+};
 
 pub(crate) mod command;
 pub(crate) mod reset;
@@ -274,7 +277,7 @@ impl Connection {
     }
 
     // Reset the device taking into account the reset after argument
-    pub fn reset_after(&mut self, is_stub: bool) -> Result<(), Error> {
+    pub fn reset_after(&mut self, is_stub: bool, chip: Chip) -> Result<(), Error> {
         let pid = self.usb_pid();
 
         match self.after_operation {
@@ -289,12 +292,62 @@ impl Connection {
                 info!("Staying in flasher stub");
                 Ok(())
             }
+            ResetAfterOperation::WatchdogReset => {
+                info!("Resetting device with watchdog");
+
+                match chip {
+                    Chip::Esp32c3 => {
+                        if self.is_using_usb_serial_jtag() {
+                            chip.into_rtc_wdt_reset()?.rtc_wdt_reset(self)?;
+                        }
+                    }
+                    Chip::Esp32p4 => {
+                        // Check if the connection is USB OTG
+                        if chip.into_usb_otg()?.is_using_usb_otg(self)? {
+                            chip.into_rtc_wdt_reset()?.rtc_wdt_reset(self)?;
+                        }
+                    }
+                    Chip::Esp32s2 => {
+                        // Check if the connection is USB OTG
+                        if chip.into_usb_otg()?.is_using_usb_otg(self)? {
+                            let target = chip.into_rtc_wdt_reset()?;
+
+                            // Check the strapping register to see if we can perform RTC WDT
+                            // reset
+                            if target.can_rtc_wdt_reset(self)? {
+                                target.rtc_wdt_reset(self)?;
+                            }
+                        }
+                    }
+                    Chip::Esp32s3 => {
+                        if self.is_using_usb_serial_jtag()
+                            || chip.into_usb_otg()?.is_using_usb_otg(self)?
+                        {
+                            let target = chip.into_rtc_wdt_reset()?;
+
+                            // Check the strapping register to see if we can perform RTC WDT
+                            // reset
+                            if target.can_rtc_wdt_reset(self)? {
+                                target.rtc_wdt_reset(self)?;
+                            }
+                        }
+                    }
+                    _ => {
+                        return Err(Error::UnsupportedFeature {
+                            chip,
+                            feature: "watchdog reset".into(),
+                        })
+                    }
+                }
+
+                Ok(())
+            }
         }
     }
 
     // Reset the device to flash mode
     pub fn reset_to_flash(&mut self, extra_delay: bool) -> Result<(), Error> {
-        if self.port_info.pid == USB_SERIAL_JTAG_PID {
+        if self.is_using_usb_serial_jtag() {
             UsbJtagSerialReset.reset(&mut self.serial)
         } else {
             #[cfg(unix)]