+++ espflash/src/targets/esp32s2.rs
@@ -3,7 +3,9 @@ use std::ops::Range;
 use xmas_elf::ElfFile;
 
 #[cfg(feature = "serialport")]
-use crate::{connection::Connection, flasher::FLASH_WRITE_SIZE, targets::MAX_RAM_BLOCK_SIZE};
+use super::flash_target::MAX_RAM_BLOCK_SIZE;
+#[cfg(feature = "serialport")]
+use crate::{connection::Connection, flasher::FLASH_WRITE_SIZE};
 use crate::{
     flasher::{FlashData, FlashFrequency},
     image_format::IdfBootloaderFormat,
@@ -34,15 +36,6 @@ const PARAMS: Esp32Params = Esp32Params::new(
 pub struct Esp32s2;
 
 impl Esp32s2 {
-    #[cfg(feature = "serialport")]
-    /// Return if the connection is USB OTG
-    fn connection_is_usb_otg(&self, connection: &mut Connection) -> Result<bool, Error> {
-        const UARTDEV_BUF_NO: u32 = 0x3fff_fd14; // Address which indicates OTG in use
-        const UARTDEV_BUF_NO_USB_OTG: u32 = 2; // Value of UARTDEV_BUF_NO when OTG is in use
-
-        Ok(connection.read_reg(UARTDEV_BUF_NO)? == UARTDEV_BUF_NO_USB_OTG)
-    }
-
     #[cfg(feature = "serialport")]
     /// Return the block2 version based on eFuses
     fn get_block2_version(&self, connection: &mut Connection) -> Result<u32, Error> {
@@ -139,7 +132,9 @@ impl Target for Esp32s2 {
 
     #[cfg(feature = "serialport")]
     fn flash_write_size(&self, connection: &mut Connection) -> Result<usize, Error> {
-        Ok(if self.connection_is_usb_otg(connection)? {
+        use super::UsbOtg;
+
+        Ok(if self.is_using_usb_otg(connection)? {
             MAX_USB_BLOCK_SIZE
         } else {
             FLASH_WRITE_SIZE
@@ -165,7 +160,9 @@ impl Target for Esp32s2 {
 
     #[cfg(feature = "serialport")]
     fn max_ram_block_size(&self, connection: &mut Connection) -> Result<usize, Error> {
-        Ok(if self.connection_is_usb_otg(connection)? {
+        use super::UsbOtg;
+
+        Ok(if self.is_using_usb_otg(connection)? {
             MAX_USB_BLOCK_SIZE
         } else {
             MAX_RAM_BLOCK_SIZE
@@ -186,5 +183,43 @@ impl Target for Esp32s2 {
 
     fn supported_build_targets(&self) -> &[&str] {
         &["xtensa-esp32s2-none-elf", "xtensa-esp32s2-espidf"]
+    }
+}
+
+#[cfg(feature = "serialport")]
+impl super::RtcWdtReset for Esp32s2 {
+    fn wdt_wprotect(&self) -> u32 {
+        0x3F40_80AC
+    }
+
+    fn wdt_config0(&self) -> u32 {
+        0x3F40_8094
+    }
+
+    fn wdt_config1(&self) -> u32 {
+        0x3F40_8098
+    }
+
+    fn can_rtc_wdt_reset(&self, connection: &mut Connection) -> Result<bool, Error> {
+        const GPIO_STRAP: u32 = 0x3F40_4038;
+        const OPTION1: u32 = 0x3F40_8128;
+        const GPIO_STRAP_SPI_BOOT_MASK: u32 = 1 << 3;
+        const FORCE_DOWNLOAD_BOOT_MASK: u32 = 0x1;
+
+        Ok(
+            connection.read_reg(GPIO_STRAP)? & GPIO_STRAP_SPI_BOOT_MASK == 0 // GPIO0 low
+                && connection.read_reg(OPTION1)? & FORCE_DOWNLOAD_BOOT_MASK == 0,
+        )
+    }
+}
+
+#[cfg(feature = "serialport")]
+impl super::UsbOtg for Esp32s2 {
+    fn uartdev_buf_no(&self) -> u32 {
+        0x3FFF_FD14
+    }
+
+    fn uartdev_buf_no_usb_otg(&self) -> u32 {
+        2
     }
 }